Checkpoint: Deletado as classes de bkp e desfeita as alterações necessarias para codar
