+++ mobile/lib/pages/login.dart
@@ -1,6 +1,12 @@
+import 'dart:convert';
+import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:wallet_mobile/dto/dto_aluno_login.dart';
 import 'package:wallet_mobile/pages/carteirinha.dart';
 import 'package:wallet_mobile/pages/cadastro.dart';
+import 'package:wallet_mobile/service/aluno_service.dart';
+import 'package:wallet_mobile/widgets/service/biometric_service.dart';
 import '/components/footer.dart';
 
 class LoginPage extends StatefulWidget {
@@ -8,17 +14,24 @@ class LoginPage extends StatefulWidget {
   _LoginPageState createState() => _LoginPageState();
 }
 
+// Adiciona biometria
 class _LoginPageState extends State<LoginPage> {
+  bool _rememberMe = false;
   bool _passwordVisible = false;
   final TextEditingController _userPasswordController = TextEditingController();
   final TextEditingController _raController = TextEditingController();
   final FocusNode _passwordFocusNode = FocusNode();
+  late BiometricService biometricService;
+  bool isAuthenticated = false;
 
   @override
   void initState() {
     super.initState();
+    biometricService = BiometricService();
+    
     _passwordFocusNode.addListener(() {
       setState(() {
+        // Mostra o ícone apenas quando o campo de senha está em foco
         _passwordVisible = _passwordFocusNode.hasFocus ? false : _passwordVisible;
       });
     });
@@ -32,16 +45,119 @@ class _LoginPageState extends State<LoginPage> {
     super.dispose();
   }
 
+  Future<void> authenticateUser() async {
+    isAuthenticated = await biometricService.authenticate();
+    if (isAuthenticated) {
+      Navigator.of(context).pushReplacement(
+        MaterialPageRoute(builder: (context) => CarteirinhaPage()),
+      );
+    } 
+  }
+
+  
+
+  Future<DtoalunoLogin> solicitarValidacaoCarteirinha() async {
+  
+    var url = Uri.parse('http://192.168.80.215:8080/cronos/rest/service/solicitacao-carteirinha/validada');
+    var response = await http.get(
+      url,
+      headers: {'Content-Type': 'application/json; charset=UTF-8'},
+    );
+
+    if (response.statusCode == 200) {
+      Map<String, dynamic> jsonData = jsonDecode(response.body);
+      DtoalunoLogin alunoDto = DtoalunoLogin.fromJson(jsonData);
+      return alunoDto;
+    } else {
+      throw Exception('Falha ao validar carteirinha: ${response.statusCode}');
+    }
+  }
+
+  Future<bool> loginCronos() async {
+  
+    var url = Uri.parse('http://192.168.80.215:8080/cronos/rest/service/login');
+    var response = await http.post(
+      url,
+      headers: {'Content-Type': 'application/json; charset=UTF-8'},
+      body: jsonEncode({
+        'ra': _raController.text,
+        'senha': _userPasswordController.text,
+      }),
+    );
+
+    if(response.statusCode == 200){
+     return true;
+    }else{
+      return false;
+    }
+  }
+
+
   bool _isFormValid() {
     return _raController.text.isNotEmpty && _userPasswordController.text.isNotEmpty;
   }
 
   Future<void> enter() async {
     if (_isFormValid()) {
-      Navigator.of(context).pushReplacement(
-        MaterialPageRoute(builder: (context) => CarteirinhaPage()),
-      );
-    }
+        bool eLoginValido = await loginCronos();
+        print("resultado login: $eLoginValido");
+        if(eLoginValido){
+            Future<DtoalunoLogin?> dadosAlunoFuture = AlunoService.recuperarAlunoSalvo();
+            DtoalunoLogin? dadosAluno = await dadosAlunoFuture; 
+
+            if (dadosAluno != null) {
+              Navigator.of(context).pushReplacement(
+                MaterialPageRoute(builder: (context) => CarteirinhaPage()),
+              );
+              return; 
+            }else{
+              DtoalunoLogin dadosAlunoRecebidoDoBackend  = await solicitarValidacaoCarteirinha();
+              if(dadosAlunoRecebidoDoBackend != null){
+                AlunoService.salvarAluno(dadosAlunoRecebidoDoBackend);
+                Navigator.of(context).pushReplacement(
+                  MaterialPageRoute(builder: (context) => CarteirinhaPage()),
+                );
+
+              }else{
+                showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: Text('Carteirinha não validada'),
+                      content: Text('Sua carteirinha ainda não foi validada. Por favor, aguarde a liberação da secretaria.'),
+                      actions: [
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                          },
+                          child: Text('OK'),
+                        ),
+                      ],
+                    );
+                  },
+                );
+            }
+          }
+        }else{
+          showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: Text('Login inválido'),
+                      content: Text('Sua senha ou ra está incorreto, lembre-se que são os mesmos dados utilizados para acessar o cronos!'),
+                      actions: [
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop(); 
+                          },
+                          child: Text('OK'),
+                        ),
+                      ],
+                    );
+                  },
+                );
+        }     
+      }
   }
 
   @override
@@ -63,7 +179,7 @@ class _LoginPageState extends State<LoginPage> {
                     child: Image.asset("assets/app/ifprlogo.png"),
                   ),
                   SizedBox(height: 20),
-                  // RA Input
+                  // Padding RA
                   Padding(
                     padding: EdgeInsets.symmetric(horizontal: 40),
                     child: Text(
@@ -75,6 +191,7 @@ class _LoginPageState extends State<LoginPage> {
                       textAlign: TextAlign.left,
                     ),
                   ),
+                  // Padding RA
                   Padding(
                     padding: EdgeInsets.symmetric(horizontal: 40),
                     child: TextFormField(
@@ -95,12 +212,14 @@ class _LoginPageState extends State<LoginPage> {
                           borderSide: BorderSide.none,
                         ),
                       ),
-                      style: TextStyle(fontSize: 20),
+                      style: TextStyle(
+                        fontSize: 20,
+                      ),
                       onChanged: (_) => setState(() {}),
                     ),
                   ),
                   SizedBox(height: 30),
-                  // Senha Input
+                  // Padding Senha
                   Padding(
                     padding: EdgeInsets.symmetric(horizontal: 40),
                     child: Text(
@@ -111,7 +230,8 @@ class _LoginPageState extends State<LoginPage> {
                       ),
                       textAlign: TextAlign.left,
                     ),
-                  ),
+                  ),                 
+                  // Padding Senha
                   Padding(
                     padding: EdgeInsets.symmetric(horizontal: 40),
                     child: TextFormField(
@@ -152,21 +272,46 @@ class _LoginPageState extends State<LoginPage> {
                               )
                             : null,
                       ),
-                      style: TextStyle(fontSize: 20),
+                      style: TextStyle(
+                        fontSize: 20,
+                      ),
                       onChanged: (_) => setState(() {}),
                     ),
                   ),
                   SizedBox(height: 20),
-                  // Botão Acessar
+                  Padding(
+                    padding: EdgeInsets.symmetric(horizontal: 40),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        Text(
+                          "Login com biometria",
+                          style: TextStyle(fontSize: 16),
+                        ),
+                        Switch(
+                          value: _rememberMe,
+                          onChanged: (bool value) {
+                            setState(() {
+                              _rememberMe = value;
+                              if(_rememberMe){
+                                biometricService.autenticar();
+                                authenticateUser();
+                              }
+                            });
+                          },
+                        ),
+                      ],
+                    ),
+                  ),
+                  SizedBox(height: 20),
+                  // Padding Acessar
                   Padding(
                     padding: EdgeInsets.symmetric(horizontal: 40),
                     child: Container(
                       height: 60,
                       alignment: Alignment.center,
                       decoration: BoxDecoration(
-                        color: _isFormValid()
-                            ? Color.fromARGB(255, 18, 129, 68)
-                            : Color.fromARGB(135, 84, 118, 99),
+                        color: _isFormValid() ? Color.fromARGB(255, 18, 129, 68) : Color.fromARGB(135, 84, 118, 99),
                         borderRadius: BorderRadius.all(
                           Radius.circular(15),
                         ),
@@ -179,11 +324,12 @@ class _LoginPageState extends State<LoginPage> {
                             fontSize: 20,
                           ),
                         ),
-                        onPressed: enter,
+                        onPressed: (){enter();},
+                        
                       ),
                     ),
                   ),
-                  // Botão Cadastrar nova senha
+                  // Padding Recuperar Senha
                   Padding(
                     padding: EdgeInsets.symmetric(horizontal: 40),
                     child: Container(
@@ -198,11 +344,11 @@ class _LoginPageState extends State<LoginPage> {
                         },
                       ),
                     ),
-                  ),
+                  ),                  
                 ],
               ),
             ),
-            Footer(),
+            Footer(), // Adiciona o Footer aqui
           ],
         ),
       ),
